requirements.txt
cryptography>=3.1
Pillow>=10.0.0
pyttsx3>=2.90
gTTS>=2.3.1
flask>=2.0.0
flask-cors>=3.0.0
pypdf>=3.0.0
langchain>=0.1.0
langchain-google-genai>=0.0.1
python-dotenv>=0.19.0
requests>=2.26.0
azure-cognitiveservices-speech>=1.24.0
gunicorn>=20.1.0
google-generativeai>=0.3.0  # Keep this version
googletrans==3.1.0a0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.2
